Fix TM2 duplicate events
Fixes duplicate events from the Peer and Stat manager both calling
CombineCrStates with their own maps of whether the event had been
generated.

Fixes the issue by creating a new StateCombiner goroutine, with a
single map. This avoids a race condition with the other solution, a
shared threadsafe map.

Divides stat combination into separate functions for Cache and
Delivery Service state, for each cache being combined. This isn't
simply for readability, but by having separate functions, it gives the
Go scheduler a chance to schedule other goroutines, preventing this
combiner goroutine (which takes ~100ms to run) from starving other
tasks.
