+++ traffic_monitor_golang/traffic_monitor/manager/health.go
@@ -81,7 +81,6 @@ func StartHealthResultManager(
 	toData todata.TODataThreadsafe,
 	localStates peer.CRStatesThreadsafe,
 	monitorConfig TrafficMonitorConfigMapThreadsafe,
-	peerStates peer.CRStatesPeersThreadsafe,
 	combinedStates peer.CRStatesThreadsafe,
 	fetchCount threadsafe.Uint,
 	errorCount threadsafe.Uint,
@@ -98,7 +97,6 @@ func StartHealthResultManager(
 		lastHealthDurations,
 		healthHistory,
 		monitorConfig,
-		peerStates,
 		combinedStates,
 		fetchCount,
 		errorCount,
@@ -116,7 +114,6 @@ func healthResultManagerListen(
 	lastHealthDurations DurationMapThreadsafe,
 	healthHistory threadsafe.ResultHistory,
 	monitorConfig TrafficMonitorConfigMapThreadsafe,
-	peerStates peer.CRStatesPeersThreadsafe,
 	combinedStates peer.CRStatesThreadsafe,
 	fetchCount threadsafe.Uint,
 	errorCount threadsafe.Uint,
@@ -135,7 +132,6 @@ func healthResultManagerListen(
 			localStates,
 			lastHealthDurations,
 			monitorConfig,
-			peerStates,
 			combinedStates,
 			fetchCount,
 			errorCount,
@@ -182,7 +178,6 @@ func processHealthResult(
 	localStates peer.CRStatesThreadsafe,
 	lastHealthDurationsThreadsafe DurationMapThreadsafe,
 	monitorConfig TrafficMonitorConfigMapThreadsafe,
-	peerStates peer.CRStatesPeersThreadsafe,
 	combinedStates peer.CRStatesThreadsafe,
 	fetchCount threadsafe.Uint,
 	errorCount threadsafe.Uint,
